Added "ChipScanning" test
This test scans the I2C bus for MCP23017 and PCA9685 chips, if the
latter type is found it will blink LEDs in a certain pattern. The
sketch also looks out for deviations from the start up state and
flashes the LEDs fast if any - this is great for burn-in tests which
can alert if there are any problems to take a closer look at.